+++ rlhf_ppo/train.py
@@ -149,7 +149,7 @@ if __name__ == '__main__':
 
     pl_model = MyPPOTransformer(config=config,model_args=model_args,training_args=training_args,lora_args=lora_args,ppo_args=ppo_args,
                                 load_in_8bit=load_in_8bit,device_map={"": trainer.fabric.local_rank} if trainer.world_size > 1 else "auto")
-
+    pl_model.bfloat16()
     # pl_ref_model = load_ref_model('../reward/best_ckpt')
     pl_ref_model = copy.deepcopy(pl_model)
     pl_ref_model.eval().half()
